+++ lib/app_style.dart
@@ -15,6 +15,9 @@ const Color kLighterBlue = Color(0xffc1d4f9);
 const Color kDarkBlue = Color(0xff19202d);
 
 const double kBorderRadius = 16.0;
+const Color kBorderColor = Color(0xffEEEEEE);
+
+const double kPaddingHorizontal = 40.0;
 
 final kBorder = OutlineInputBorder(
   borderRadius: BorderRadius.circular(kBorderRadius),
